Revert "changing money to use bigdecimal"
This reverts commit 782a1437fb3d6f8cce0fa51f6f5552abd83e3504.

+++ src/org/capitalism/Product.java
@@ -9,7 +9,7 @@ public class Product implements IConsumable {
 	
 	public Product(Money money) {
 
-		value = money.getValue().doubleValue();
+		value = money.getValue();
 		priceMultiplier = new Random().nextDouble() + 1;
 		Money.MoneySupply.giveMoney(money);
 	}
